Split ssl misc + version tests
